+++ index.js
@@ -14,7 +14,11 @@ try {
   var EventEmitter = require('events').EventEmitter;
   if (!EventEmitter) throw new Error();
 } catch (err) {
-  var Emitter = require('emitter');
+  try {
+    var Emitter = require('emitter')
+  } catch (err) {
+    throw err
+  }
 }
 
 /**